reorganize coroutine and dispatcher modules
