--- app/src/main/java/com/jonecx/qio/di/DispatchersModule.kt
@@ -1,22 +1,0 @@
-package com.jonecx.qio.di
-
-import com.jonecx.qio.di.QioDispatchers.Default
-import com.jonecx.qio.di.QioDispatchers.IO
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-
-@Module
-@InstallIn(SingletonComponent::class)
-class DispatchersModule {
-    @Provides
-    @Dispatcher(IO)
-    fun providesIODispatcher(): CoroutineDispatcher = Dispatchers.IO
-
-    @Provides
-    @Dispatcher(Default)
-    fun providesDefaultDispatcher(): CoroutineDispatcher = Dispatchers.Default
-}
