+++ app/src/main/java/com/jonecx/qio/di/QioCoroutineModule.kt
@@ -1,0 +1,49 @@
+package com.jonecx.qio.di
+
+import com.jonecx.qio.di.QioDispatchers.Default
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+import javax.inject.Qualifier
+import javax.inject.Singleton
+
+@Qualifier
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ApplicationScope
+
+@Qualifier
+@Retention(AnnotationRetention.RUNTIME)
+annotation class Dispatcher(val qioDispatcher: QioDispatchers)
+
+enum class QioDispatchers {
+    Default,
+    IO,
+}
+
+@Module
+@InstallIn(SingletonComponent::class)
+object CoroutineScopesModule {
+    @Provides
+    @Singleton
+    @ApplicationScope
+    fun providesCoroutineScope(
+        @Dispatcher(Default) dispatcher: CoroutineDispatcher,
+    ): CoroutineScope = CoroutineScope(SupervisorJob() + dispatcher)
+}
+
+@Module
+@InstallIn(SingletonComponent::class)
+class DispatchersModule {
+    @Provides
+    @Dispatcher(QioDispatchers.IO)
+    fun providesIODispatcher(): CoroutineDispatcher = Dispatchers.IO
+
+    @Provides
+    @Dispatcher(Default)
+    fun providesDefaultDispatcher(): CoroutineDispatcher = Dispatchers.Default
+}
