--- app/src/main/java/com/jonecx/qio/di/QioDispatchers.kt
@@ -1,13 +1,0 @@
-package com.jonecx.qio.di
-
-import javax.inject.Qualifier
-import kotlin.annotation.AnnotationRetention.RUNTIME
-
-@Qualifier
-@Retention(RUNTIME)
-annotation class Dispatcher(val qioDispatcher: QioDispatchers)
-
-enum class QioDispatchers {
-    Default,
-    IO,
-}
